+++ Minor.Miffy.MicroServices/Host/HostEventArgs/EventMessageReceivedEventArgs.cs
@@ -1,0 +1,11 @@
+using System;
+using System.Collections.Generic;
+
+namespace Minor.Miffy.MicroServices.Host.HostEventArgs
+{
+    public class EventMessageReceivedEventArgs : EventArgs
+    {
+        public string QueueName { get; set; }
+        public IEnumerable<string> TopicExpressions { get; set; }
+    }
+}