+++ Minor.Miffy.MicroServices/Host/HostEventArgs/HostPausedEventArgs.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Minor.Miffy.MicroServices.Host.HostEventArgs
+{
+    public class HostPausedEventArgs : EventArgs
+    {
+    }
+}